Move License and Version command line tools to separate folders
...and create .cmd files to access them directly.

+++ License/fuGenerate.py
@@ -1,45 +1,40 @@
-import json, os, re, sys, datetime;
+import re, os;
 
-sMainFolderPath = os.path.dirname(os.path.abspath(__file__));
-sParentFolderPath = os.path.normpath(os.path.join(sMainFolderPath, ".."));
-sModulesFolderPath = os.path.join(sMainFolderPath, "modules");
-asOriginalSysPath = sys.path[:];
-sys.path = [sMainFolderPath, sParentFolderPath, sModulesFolderPath] + sys.path;
-
-from mProductVersionAndLicense.cDate import cDate;
-from mProductVersionAndLicense.cProductDetails import cProductDetails;
-from mProductVersionAndLicense.cLicenseConfiguration import cLicenseConfiguration;
-from mProductVersionAndLicense.cLicenseCheckServer import cLicenseCheckServer;
-
-# Restore the search path
-sys.path = asOriginalSysPath;
+from mProductDetails.cDate import cDate;
+from mProductDetails.cProductDetails import cProductDetails;
+from mProductDetails.cLicenseConfiguration import cLicenseConfiguration;
+from mProductDetails.cLicenseCheckServer import cLicenseCheckServer;
 
 def fUsage(sMainScriptName, sFeatureName):
          ################################################################################
   print "Generate new licenses for a specific product";
   print "Usage:";
-  print "  %s %s <options>" % (sMainScriptName, sFeatureName);
-  print "Where <options> is:";
+  print "  %s %s <settings> <options>" % (sMainScriptName, sFeatureName);
+  print "Where <settings> are:";
   print " [--product=]<product folder>      The root folder for the product.";
   print " [--config=]<config file name>     Path to a license configuration file.";
-  print " [--version=<version>]             Select a license version (optional).";
-  print "  --licensee=<licensee name>       Name of the licensee.";
-  print "  --use=<usage keyword>            Licensed usage keyword; valid values depend";
+  print " [--licensee=]<licensee name>      Name of the licensee.";
+  print " [--use=]<usage keyword>           Licensed usage keyword; valid values depend";
   print "                                   on the selected license configuration."
-  print " [--instances=<maximum>]           Maximum number of instances for the license";
+  print "And <options> are:";
+  print " [--version=<version>]             Select a license version (optional).";
+  print " [--instances=<number>]            Maximum number of instances for the license";
   print "                                   (optional; default = 1).";
-  print " [--start-date=<Y-M-D date>]       Start date (optional; default = today).";
-  print " [--end-date=<Y-M-D> date]         End date (optional).";
-  print " [--duration=<Ny/m/d duration>]    Duration (optional; default = 1y).";
+  print " [--start-date=<YY-MM-DD date>]    Start date (optional; default = today).";
+  print " [--end-date=<YY-MM-DD date>]      End date (optional).";
+  print " [--duration=<#y+#m+#d duration>]  Duration (optional; default = 1y+7d when";
+  print "                                   starting today, otherwise 1y)";
   print " [--output=<license file name>]    Where to write the generated license.";
   print "";
   print "This feature will generate a new license for given settings, check with the";
   print "server that the license is considered valid and write it to a file.";
 
-def License_fuMain_Generate(sMainScriptName, sFeatureName, asArguments):
-  if len(asArguments) == 0:
+def fuGenerate(sMainScriptName, sFeatureName, asArguments, dsArguments):
+  if len(asArguments) + len(dsArguments) == 0:
     fUsage(sMainScriptName, sFeatureName);
     return 1;
+  
+  # Parse arguments
   sProductFolderPath = None;
   sLicenseConfigurationFilePath = None;
   sLicenseVersion = None;
@@ -51,67 +46,77 @@ def License_fuMain_Generate(sMainScriptName, sFeatureName, asArguments):
   oEndDate = None;
   sDuration = None;
   sOutputFilePath = None;
-  for sArgument in asArguments:
-    oArgumentMatch = re.match(r"^\-\-([\w\-]+)(?:\=(.*))?$", sArgument);
-    if not oArgumentMatch:
-      if sProductFolderPath is None:
-        sProductFolderPath = sArgument;
-      elif sLicenseConfigurationFilePath is None:
-        sLicenseConfigurationFilePath = sArgument;
-      else:
-        print "- Superfluous argument %s!" % sArgument;
-        return 1;
+  for (sName, sValue) in dsArguments.items():
+    if sName == "help":
+      fUsage(sMainScriptName, sFeatureName);
+      return 0;
+    elif sName == "product":
+      sProductFolderPath = sValue;
+    elif sName == "config":
+      sLicenseConfigurationFilePath = sValue;
+    elif sName == "version":
+      sLicenseVersion = sValue;
+    elif sName == "licensee":
+      sLicenseeName = sValue;
+    elif sName == "use":
+      sUsageTypeKeyword = sValue;
+    elif sName == "instances":
+      uLicensedInstances = long(sValue);
+    elif sName == "start":
+      oStartDate = cDate.foFromString(sValue);
+    elif sName == "end":
+      oEndDate = cDate.foFromString(sValue);
+    elif sName == "duration":
+      sDuration = sValue;
+    elif sName == "output":
+      sOutputFilePath = sValue;
     else:
-      (sName, sValue) = oArgumentMatch.groups();
-      if sName == "help":
-        fUsage(sMainScriptName, sFeatureName);
-        return 0;
-      elif sName == "product":
-        sProductFolderPath = sValue;
-      elif sName == "config":
-        sLicenseConfigurationFilePath = sValue;
-      elif sName == "version":
-        sLicenseVersion = sValue;
-      elif sName == "licensee":
-        sLicenseeName = sValue;
-      elif sName == "use":
-        sUsageTypeKeyword = sValue;
-      elif sName == "instances":
-        uLicensedInstances = long(sValue);
-      elif sName == "start":
-        oStartDate = cDate.foFromString(sValue);
-      elif sName == "end":
-        oEndDate = cDate.foFromString(sValue);
-      elif sName == "duration":
-        sDuration = sValue;
-      elif sName == "output":
-        sOutputFilePath = sValue;
-      else:
-        print "- Unrecognized argument %s!" % sArgument;
-        return 1;
+      print "- Unrecognized argument %s!" % sArgument;
+      fUsage(sMainScriptName, sFeatureName);
+      return 1;
+  
+  for sArgument in asArguments:
+    if sProductFolderPath is None:
+      sProductFolderPath = sArgument;
+    elif sLicenseConfigurationFilePath is None:
+      sLicenseConfigurationFilePath = sArgument;
+    elif sLicenseeName is None:
+      sLicenseeName = sArgument;
+    elif sUsageTypeKeyword is None:
+      sUsageTypeKeyword = sArgument;
+    else:
+      print "- Superfluous argument %s!" % sArgument;
+      fUsage(sMainScriptName, sFeatureName);
+      return 1;
+  
   # Check arguments
   if sProductFolderPath is None:
     print "- Please provide a --product argument!";
+    fUsage(sMainScriptName, sFeatureName);
     return 1;
   if sLicenseConfigurationFilePath is None:
     print "- Please provide an --config argument!";
+    fUsage(sMainScriptName, sFeatureName);
     return 1;
   if sLicenseeName is None:
     print "- Please provide an --licensee argument!";
+    fUsage(sMainScriptName, sFeatureName);
     return 1;
   if sUsageTypeKeyword is None:
     print "- Please provide an --use argument!";
+    fUsage(sMainScriptName, sFeatureName);
     return 1;
   if oEndDate is not None and sDuration is not None:
     print "- Please provide either --end or --duration, not both!";
+    fUsage(sMainScriptName, sFeatureName);
     return 1;
-  oDurationMatch = None;
   if sDuration is not None:
-    oDurationMatch = re.match(r"(\d+)([ymd])", sDuration.lower());
-    if oDurationMatch is None:
-      print "- Please provide a valid value (<number>[ymd]) for duration!";
-      return 1;
-  
+    for sDurationComponent in sDuration.split("+"):
+      oDurationMatch = re.match(r"(\d+)([ymd])", sDurationComponent, re.I);
+      if oDurationMatch is None:
+        print "- Please provide a valid value (<number>[ymd]) for duration!";
+        fUsage(sMainScriptName, sFeatureName);
+        return 1;
   # Read the product details from the product folder:
   oProductDetails = cProductDetails.foReadFromFolderPath(sProductFolderPath);
   if not oProductDetails:
@@ -125,34 +130,38 @@ def License_fuMain_Generate(sMainScriptName, sFeatureName, asArguments):
   print "+ Read %d license configurations from %s." % (len(aoOrderedLicenseConfigurations), sLicenseConfigurationFilePath);
 
   # Filter license configurations for product name
-  aoOrderedLicenseConfigurations = [
+  aoSelectedOrderedLicenseConfigurations = [
     oLicenseConfiguration for oLicenseConfiguration in aoOrderedLicenseConfigurations
     if oProductDetails.sProductName in oLicenseConfiguration.asProductNames
   ];
-  if not aoOrderedLicenseConfigurations:
+  if not aoSelectedOrderedLicenseConfigurations:
     print "- No license configurations are available for product %s" % oProductDetails.sProductName;
     return 1;
-  print "+ Selected %d license configurations for product %s." % \
-      (len(aoOrderedLicenseConfigurations), oProductDetails.sProductName);
+  if len(aoOrderedLicenseConfigurations) != len(aoSelectedOrderedLicenseConfigurations):
+    print "+ Selected %d license configurations for product %s." % \
+        (len(aoSelectedOrderedLicenseConfigurations), oProductDetails.sProductName);
   if sLicenseVersion:
     # Filter for version
-    aoOrderedLicenseConfigurations = [
-      oLicenseConfiguration for oLicenseConfiguration in aoOrderedLicenseConfigurations
+    aoSelectedOrderedLicenseConfigurations = [
+      oLicenseConfiguration for oLicenseConfiguration in aoSelectedOrderedLicenseConfigurations
       if oLicenseConfiguration.sVersion == sLicenseVersion
     ];
-    if len(aoOrderedLicenseConfigurations) == 0:
+    if len(aoSelectedOrderedLicenseConfigurations) == 0:
       print "- No license configurations are available for product %s and version %s" % \
           (oProductDetails.sProductName, sLicenseVersion);
       return 1;
-    assert len(aoOrderedLicenseConfigurations) == 1, \
+    assert len(aoSelectedOrderedLicenseConfigurations) == 1, \
         "There are multiple license configurations available for products %s and version %s" % \
           (oProductDetails.sProductName, sLicenseVersion);
   # Select the latest
-  oLicenseConfiguration = aoOrderedLicenseConfigurations[-1];
-  print "+ Selected license configurations version %s." % (oLicenseConfiguration.sLicenseVersion);
+  oLicenseConfiguration = aoSelectedOrderedLicenseConfigurations[-1];
+  if len(aoSelectedOrderedLicenseConfigurations) > 1:
+    print "+ Selected license configurations version %s." % (oLicenseConfiguration.sLicenseVersion);
   # Apply start/end dates
   if oEndDate is not None:
     pass;
+  elif oStartDate == cDate.foNow():
+    oEndDate = oStartDate.foEndDateForDuration(sDuration or "1y+7d");
   else:
     oEndDate = oStartDate.foEndDateForDuration(sDuration or "1y");
   # Create a new license.
@@ -177,6 +186,10 @@ def License_fuMain_Generate(sMainScriptName, sFeatureName, asArguments):
     print "- The server reported the new generated license as having exceeded the licensed number of instances!";
     return 2;
   if sOutputFilePath:
-    open(sOutputFilePath, "wb").write(oLicense.sLicenseBlock);
-    print "+ The new license has been written to file.";
+    if os.path.isfile(sOutputFilePath):
+      open(sOutputFilePath, "ab").write(oLicense.sLicenseBlock);
+      print "+ The new license has been appended to %s." % sOutputFilePath;
+    else:
+      open(sOutputFilePath, "wb").write(oLicense.sLicenseBlock);
+      print "+ The new license has been written to %s." % sOutputFilePath;
   return 0;
